Revert CI cleanup build service
We found that each build service takes ~14M extra memory, 100
subprojects mean 1.4G memory, that's too much. Before we find a
solution, let's revert this change.
